Домашнее задание № 7, MVP
