+++ app/src/main/java/ru/androidschool/intensiv/presentation/feed/MainCardContainer.kt
@@ -1,4 +1,4 @@
-package ru.androidschool.intensiv.ui.feed
+package ru.androidschool.intensiv.presentation.feed
 
 import android.view.View
 import androidx.annotation.StringRes
