+++ app/src/main/res/layout/feed_fragment.xml
@@ -19,7 +19,7 @@
             android:layout_height="wrap_content"
             android:orientation="vertical">
 
-            <ru.androidschool.intensiv.ui.custom.MovieProgressFormView
+            <ru.androidschool.intensiv.presentation.custom.MovieProgressFormView
                 android:id="@+id/progress_view"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
