+++ app/src/main/res/layout/fragment_profile.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorPrimary"
-    tools:context=".ui.profile.ProfileFragment">
+    tools:context=".presentation.profile.ProfileFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
