+++ app/src/main/res/layout/fragment_tv_shows.xml
@@ -5,10 +5,10 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorPrimary"
-    tools:context="ru.androidschool.intensiv.ui.tvshows.TvShowsFragment"
+    tools:context="ru.androidschool.intensiv.presentation.tvshows.TvShowsFragment"
     android:orientation="vertical">
 
-    <ru.androidschool.intensiv.ui.custom.MovieProgressFormView
+    <ru.androidschool.intensiv.presentation.custom.MovieProgressFormView
         android:id="@+id/progress_view"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
