+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -5,7 +5,7 @@
     app:startDestination="@+id/home_dest">
     <fragment
         android:id="@+id/home_dest"
-        android:name="ru.androidschool.intensiv.ui.feed.FeedFragment"
+        android:name="ru.androidschool.intensiv.presentation.feed.FeedFragment"
         android:label="@string/home"
         tools:layout="@layout/feed_fragment">
         <action
@@ -18,7 +18,7 @@
 
     <fragment
         android:id="@+id/tv_shows_dest"
-        android:name="ru.androidschool.intensiv.ui.tvshows.TvShowsFragment"
+        android:name="ru.androidschool.intensiv.presentation.tvshows.TvShowsFragment"
         android:label="@string/tv_show"
         tools:layout="@layout/fragment_tv_shows">
 
@@ -29,7 +29,7 @@
 
     <fragment
         android:id="@+id/movie_details_fragment"
-        android:name="ru.androidschool.intensiv.ui.movie_details.MovieDetailsFragment"
+        android:name="ru.androidschool.intensiv.presentation.movie_details.MovieDetailsFragment"
         android:label="fragment_details"
         tools:layout="@layout/fragment_movie_details" >
         <argument
@@ -39,13 +39,13 @@
 
     <fragment
         android:id="@+id/profile_fragment"
-        android:name="ru.androidschool.intensiv.ui.profile.ProfileFragment"
+        android:name="ru.androidschool.intensiv.presentation.profile.ProfileFragment"
         android:label="fragment_profile"
         tools:layout="@layout/fragment_profile" />
 
     <fragment
         android:id="@+id/search_dest"
-        android:name="ru.androidschool.intensiv.ui.search.SearchFragment"
+        android:name="ru.androidschool.intensiv.presentation.search.SearchFragment"
         android:label="fragment_search"
         tools:layout="@layout/fragment_search" >
         <argument
